+++ src/Access/Ilias.php
@@ -2,6 +2,7 @@
 
 namespace srag\Plugins\SrTile\Access;
 
+use ilObject;
 use ilObjUser;
 use ilSrTilePlugin;
 use srag\DIC\SrTile\DICTrait;
@@ -9,6 +10,7 @@ use srag\Plugins\SrTile\Certificate\Certificates;
 use srag\Plugins\SrTile\Favorite\Favorites;
 use srag\Plugins\SrTile\LearningProgress\LearningProgress;
 use srag\Plugins\SrTile\LearningProgress\LearningProgressBar;
+use srag\Plugins\SrTile\Metadata\Metadata;
 use srag\Plugins\SrTile\Utils\SrTileTrait;
 
 /**
@@ -26,14 +28,14 @@ final class Ilias {
 	/**
 	 * @var self
 	 */
-	protected static $instance = NULL;
+	protected static $instance = null;
 
 
 	/**
 	 * @return self
 	 */
 	public static function getInstance(): self {
-		if (self::$instance === NULL) {
+		if (self::$instance === null) {
 			self::$instance = new self();
 		}
 
@@ -95,5 +97,15 @@ final class Ilias {
 	 */
 	public function learningProgressBar(ilObjUser $user, int $obj_ref_id): LearningProgressBar {
 		return LearningProgressBar::getInstance($user, $obj_ref_id);
+	}
+
+
+	/**
+	 * @param ilObject $il_object
+	 *
+	 * @return Metadata
+	 */
+	public function metadata(ilObject $il_object): Metadata {
+		return Metadata::getInstance($il_object);
 	}
 }