+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [x]
 - Make possible to style spec. tile icons
 - Allows incomplete configuration if the tile is disabled
+- Show language flag
 
 ## [1.4.3]
 - Display placeholder types in notification form